fix: taking the required functions from `latticeExtra` to resolve "dependency hell"
`ggplot2` and `lattice` was added as 'Suggested' packages to DESCRIPTION file -
so if a user would use them, we can rely on those helpers. To unify `lattice` images
the was calling some `latticeExtra` functions too, but it might be possible that a user
has not installed `latticeExtra` which would lead to some errors. It's not possible (AFAIK)
to warn user about this without adding all to 'Depend'/'Imports', so I rather copied/modified
the required 3 functions from `latticeExtra` with references to them.