row.names and col.names argument for pandoc.table()
I think these arguments are important, therefore I propose to place
them early in the argument list. This is a matter of opinion of course.
I tried to use match.arg on these, but it throws an error. Maybe the
arguments need to be renamed for this to work, but I think it's good to
be compatible to knitr::kable() regarding the naming of the arguments.

I also tried to place the code later in the function definition of
pandoc.table, where the other arguments are initialized, but then it
fails when row.names is set to FALSE, so it seems some code uses
the row names before that place.

Closes #281.